add exception for TaskService and StateService

+++ src/main/java/com/softserve/itacademy/service/impl/StateServiceImpl.java
@@ -1,17 +1,19 @@
 package com.softserve.itacademy.service.impl;
 
+import com.softserve.itacademy.exception.NullEntityReferenceException;
 import com.softserve.itacademy.model.State;
 import com.softserve.itacademy.repository.StateRepository;
 import com.softserve.itacademy.service.StateService;
 import org.springframework.stereotype.Service;
 
+import javax.persistence.EntityNotFoundException;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
 @Service
 public class StateServiceImpl implements StateService {
-    private StateRepository stateRepository;
+    private final StateRepository stateRepository;
 
     public StateServiceImpl(StateRepository stateRepository) {
         this.stateRepository = stateRepository;
@@ -19,31 +21,48 @@ public class StateServiceImpl implements StateService {
 
     @Override
     public State create(State state) {
+        try {
             return stateRepository.save(state);
+        } catch (RuntimeException e) {
+            throw new NullEntityReferenceException("State can't be 'null'");
+        }
     }
 
     @Override
     public State readById(long id) {
         Optional<State> optional = stateRepository.findById(id);
+        if (optional.isPresent()) {
             return optional.get();
+        }
+        throw new EntityNotFoundException("State with id " + id + " not found");
     }
 
     @Override
     public State update(State state) {
-            State oldState = readById(state.getId());
+        if (state != null) {
+            if (readById(state.getId()) != null) {
                 return stateRepository.save(state);
+            }
+        }
+        throw new NullEntityReferenceException("State can't be 'null'");
     }
 
     @Override
     public void delete(long id) {
         State state = readById(id);
-            stateRepository.delete(state);
+        if (state == null) {
+            throw new EntityNotFoundException("State with id " + id + " not found");
+        }
+        stateRepository.delete(state);
     }
 
     @Override
     public State getByName(String name) {
         Optional<State> optional = Optional.ofNullable(stateRepository.getByName(name));
+        if (optional.isPresent()) {
             return optional.get();
+        }
+        throw new EntityNotFoundException("State with name " + name + " not found");
     }
 
     @Override
